backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
supabase>=2.5.0
groq>=0.9.0
google-generativeai>=0.7.0
python-jose[cryptography]>=3.3.0
pydantic>=2.6.0
httpx>=0.27.0
python-multipart>=0.0.9
